+++ packages/gamut-labs/src/LayoutMenu/LayoutMenu.tsx
@@ -35,9 +35,9 @@ export const LayoutMenu: React.FC<LayoutMenuProps> = ({
   breakpoint = 'lg',
   children,
 }) => {
-  const closeFlyoutRef = useRef(() => {});
+  const closeFlyoutRef = useRef<() => void>();
 
-  const closeFlyout = () => closeFlyoutRef.current();
+  const closeFlyout = () => closeFlyoutRef.current?.();
 
   const accordionMenuSections = sections.map((section) => (
     <AccordionMenu