+++ packages/gamut-kit/CHANGELOG.md
@@ -3,6 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.5.82](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.81...@codecademy/gamut-kit@0.5.82) (2021-09-16)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.81](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.80...@codecademy/gamut-kit@0.5.81) (2021-09-16)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.80](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.79...@codecademy/gamut-kit@0.5.80) (2021-09-15)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.79](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.78...@codecademy/gamut-kit@0.5.79) (2021-09-15)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.78](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.77...@codecademy/gamut-kit@0.5.78) (2021-09-15)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.77](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.76...@codecademy/gamut-kit@0.5.77) (2021-09-14)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.76](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.75...@codecademy/gamut-kit@0.5.76) (2021-09-14)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.75](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.74...@codecademy/gamut-kit@0.5.75) (2021-09-14)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.74](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.73...@codecademy/gamut-kit@0.5.74) (2021-09-13)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.73](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.72...@codecademy/gamut-kit@0.5.73) (2021-09-13)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.72](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.71...@codecademy/gamut-kit@0.5.72) (2021-09-10)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
+### [0.5.71](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.70...@codecademy/gamut-kit@0.5.71) (2021-09-10)
+
+**Note:** Version bump only for package @codecademy/gamut-kit
+
+
+
+
+
 ### [0.5.70](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-kit@0.5.69...@codecademy/gamut-kit@0.5.70) (2021-09-09)
 
 **Note:** Version bump only for package @codecademy/gamut-kit