+++ packages/gamut-patterns/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.6.1](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-patterns@0.6.0...@codecademy/gamut-patterns@0.6.1) (2021-09-13)
+
+**Note:** Version bump only for package @codecademy/gamut-patterns
+
+
+
+
+
+## [0.6.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-patterns@0.5.8...@codecademy/gamut-patterns@0.6.0) (2021-09-10)
+
+
+### Features
+
+* bump ESLint configs to latest ([e0a201a](https://github.com/Codecademy/client-modules/commit/e0a201abcc5f49718538d3d91af21cb37db4470e))
+
+
+
 ### [0.5.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-patterns@0.5.7...@codecademy/gamut-patterns@0.5.8) (2021-08-30)
 
 **Note:** Version bump only for package @codecademy/gamut-patterns